+++ src/site/pages/Home.tsx
@@ -5,6 +5,7 @@ import { SectionHeading } from "../components/SectionHeading";
 import { Logo } from "../components/Logo";
 import { HeroBackground } from "../components/HeroBackground";
 import { useDesign } from "../design/DesignContext";
+import { heroById } from "../data/heroVariants";
 import { Link } from "../router";
 import { TAGLINE } from "../data/site";
 
@@ -18,7 +19,8 @@ const FEATURES = [
 ];
 
 export function Home() {
-  const { heroId } = useDesign();
+  const { choice } = useDesign();
+  const heroId = heroById(choice("hero")).id;
 
   return (
     <>